% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLoadIndex.R
\name{GetLoadIndex}
\alias{GetLoadIndex}
\title{Scale active ingredient indicator values to be 0-1}
\usage{
GetLoadIndex(f_dat = pli_exdat)
}
\arguments{
\item{f_dat}{Active ingredient information, see pli_examdat}
}
\value{
A dataframe with the indicator, the active ingredient, the original value, and the index value
}
\description{
Scale active ingredient indicator values to be 0-1
}
