% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pli_exdat}
\alias{pli_exdat}
\title{Example cleaned data from an elicitation}
\format{
A data frame with an expanding number of rows and 4 variables:
\describe{
\item{title}{The title of the elicitation}
\item{scenario}{Must be either CCP (Current Commercial Practice) or IPM (Integrated Pest Management)}
\item{ai_name}{Active ingredient (ai) name, must match what is in the online PPDB search}
\item{ai_con_g_g}{The concentration of the ai in the product used (grams ai per gram of product)}
\item{prod_amt_g_ha}{Amount of product applied in grams per hectare}
}
}
\usage{
pli_exdat
}
\description{
A dataset containing the needed information to extract Load Indices
}
\keyword{datasets}
