% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pli_exindices}
\alias{pli_exindices}
\title{Example of calculated indices}
\format{
A data frame with an expanding number of rows and 4 variables:
\describe{
\item{pli_cat}{Either env_tox_load (environmental toxicity load) or env_fate_load (environmental fate load)}
\item{name}{Name of one of the 14 indicators from the PPDB}
\item{substance}{the active ingredient (ai) being evaluated}
\item{value.num}{the value of the ai in a given indicator, in the units of that indicator}
\item{load_index}{the value scaled (using a logarithmic model) to 0-1}
\item{load_index2}{the load_index, adjusted so low is 'good' and high is 'bad'}
}
}
\usage{
pli_exindices
}
\description{
A dataset containing an example of output from the GetLoadIndex function
}
\keyword{datasets}
